move globalEventBus to eventBus.js
