+++ js/eventBus.js
@@ -38,4 +38,7 @@ class EventBus {
 
 }
 
+
+export let globalEventBus = new EventBus();
+
 export default EventBus;
