+++ js/sceneManager.js
@@ -1,5 +1,5 @@
 import * as THREE from "./threejs_es6/three.module.js";
-import { globalEventBus } from "./main.js";
+import { globalEventBus } from "./eventBus.js";
 import CubicMapManager from "./cubicMapManager.js";
 import ColorRampPalette from "./colorRampPalette.js";
 import OrbitalCamera from "./orbitalCamera.js";
